Move :parser:core into sdk
