+++ components/parser/kmp/svg/build.gradle.kts
@@ -8,7 +8,7 @@ kotlin {
     sourceSets {
         commonMain.dependencies {
             implementation(projects.sdk.ir.core)
-            implementation(projects.components.parser.common)
+            implementation(projects.sdk.parser.common)
             implementation(libs.xmlutil)
         }
         commonTest.dependencies {
