+++ components/parser/kmp/svg/src/commonMain/kotlin/io/github/composegears/valkyrie/parser/kmp/svg/SVGParser.kt
@@ -1,6 +1,5 @@
 package io.github.composegears.valkyrie.parser.kmp.svg
 
-import io.github.composegears.valkyrie.parser.common.PathParser
 import io.github.composegears.valkyrie.sdk.ir.core.IrColor
 import io.github.composegears.valkyrie.sdk.ir.core.IrFill
 import io.github.composegears.valkyrie.sdk.ir.core.IrImageVector
@@ -8,6 +7,7 @@ import io.github.composegears.valkyrie.sdk.ir.core.IrPathFillType
 import io.github.composegears.valkyrie.sdk.ir.core.IrPathNode
 import io.github.composegears.valkyrie.sdk.ir.core.IrStroke
 import io.github.composegears.valkyrie.sdk.ir.core.IrVectorNode
+import io.github.composegears.valkyrie.sdk.parser.common.PathParser
 
 private typealias ViewBox = List<Float>
