+++ sdk/parser/common/src/commonMain/kotlin/io/github/composegears/valkyrie/sdk/parser/common/PathNode.kt
@@ -1,4 +1,4 @@
-package io.github.composegears.valkyrie.parser.common
+package io.github.composegears.valkyrie.sdk.parser.common
 
 import io.github.composegears.valkyrie.sdk.ir.core.IrPathNode
 import io.github.composegears.valkyrie.sdk.ir.core.IrPathNode.ArcTo
