+++ sdk/parser/common/src/commonTest/kotlin/io/github/composegears/valkyrie/sdk/parser/common/AndroidColorParserTest.kt
@@ -1,4 +1,4 @@
-package io.github.composegears.valkyrie.parser.common
+package io.github.composegears.valkyrie.sdk.parser.common
 
 import assertk.assertThat
 import assertk.assertions.isEqualTo
