+++ settings.gradle.kts
@@ -75,7 +75,6 @@ include("components:generator:kmp:imagevector")
 include("components:generator:iconpack")
 include("components:generator:jvm:poet-extensions")
 include("components:generator:jvm:imagevector")
-include("components:parser:common")
 include("components:parser:jvm:svg")
 include("components:parser:jvm:xml")
 include("components:parser:kmp:xml")
@@ -97,6 +96,7 @@ include("sdk:ir:compose")
 include("sdk:ir:core")
 include("sdk:ir:util")
 include("sdk:ir:xml")
+include("sdk:parser:common")
 include("sdk:shared")
 include("sdk:test:coverage")
 include("sdk:test:resource-loader")
